+++ testing/ubuntu24.10/test.sh
@@ -11,6 +11,7 @@ clean=0
 cleanOnly=0
 release=0
 debug=0
+installOnly=0
 
 for arg in $@
 do
@@ -29,7 +30,16 @@ do
     if [[ $arg = "--debug" ]]; then
         debug=1
     fi
+
+    if [[ $arg = "--install-only" ]]; then
+        installOnly=1
+    fi
 done
+
+if [[ $release -eq 0 ]] && [[ $debug -eq 0 ]]; then
+    echo "No build type selected, please specify either --debug, --release or both."
+    exit 1
+fi
 
 if [[ $clean -eq 1 ]] || [[ $cleanOnly -eq 1 ]]; then
     distrobox rm -f ${distroboxName}
@@ -40,16 +50,11 @@ if [[ $cleanOnly -eq 1 ]]; then
     exit 0
 fi
 
-if [[ $release -eq 0 ]] && [[ $debug -eq 0 ]]; then
-    echo "No build type selected, please specify either --debug, --release or both."
-    exit 1
-fi
-
 if [ ! -d "home" ]; then
     mkdir home
     distrobox create -n -Y ${distroboxName} -i ${distroboxImage} --home ${distroboxHomeDir}
 
-    distrobox enter ${distroboxName} -nw -T -e sudo apt install -y okular
+    distrobox enter ${distroboxName} -nw -T -e sudo dnf install -y okular
 fi
 
 if [[ $debug -eq 1 ]]; then
@@ -71,10 +76,12 @@ if [[ $debug -eq 1 ]]; then
 
     distrobox enter ${distroboxName} --no-workdir -T -e sudo ./install.sh
 
-    echo "Testing the v3d plugin on Okular version "${okularVersion}" on "${distro}" using teapot.v3d and debug plugins"
-    distrobox enter ${distroboxName} --no-workdir -T -e okular examples/teapot.v3d
-    echo "Testing the pdf plugin on Okular version "${okularVersion}" on "${distro}" using modelGrid.pdf and debug plugins"
-    distrobox enter ${distroboxName} --no-workdir -T -e okular examples/modelGrid.pdf
+    if [[ $installOnly -eq 0 ]]; then
+        echo "Testing the v3d plugin on Okular version "${okularVersion}" on "${distro}" using teapot.v3d and debug plugins"
+        distrobox enter ${distroboxName} --no-workdir -T -e okular examples/teapot.v3d
+        echo "Testing the pdf plugin on Okular version "${okularVersion}" on "${distro}" using modelGrid.pdf and debug plugins"
+        distrobox enter ${distroboxName} --no-workdir -T -e okular examples/modelGrid.pdf
+    fi
 fi
 
 if [[ $release -eq 1 ]]; then
@@ -96,8 +103,11 @@ if [[ $release -eq 1 ]]; then
 
     distrobox enter ${distroboxName} --no-workdir -T -e sudo ./install.sh
 
-    echo "Testing the v3d plugin on Okular version "${okularVersion}" on "${distro}" using teapot.v3d and release plugins"
-    distrobox enter ${distroboxName} --no-workdir -T -e okular examples/teapot.v3d
-    echo "Testing the pdf plugin on Okular version "${okularVersion}" on "${distro}" using modelGrid.pdf and release plugins"
-    distrobox enter ${distroboxName} --no-workdir -T -e okular examples/modelGrid.pdf
+    if [[ $installOnly -eq 0 ]]; then
+        echo "Testing the v3d plugin on Okular version "${okularVersion}" on "${distro}" using teapot.v3d and release plugins"
+        distrobox enter ${distroboxName} --no-workdir -T -e okular examples/teapot.v3d
+        echo "Testing the pdf plugin on Okular version "${okularVersion}" on "${distro}" using modelGrid.pdf and release plugins"
+        distrobox enter ${distroboxName} --no-workdir -T -e okular examples/modelGrid.pdf
+    fi
 fi
+
